requirements.txt
tensorflow-cpu==2.16.1
keras==2.16.0rc0  # Compatible pre-release version
scikit-learn==1.3.2
alibi-detect==0.12.0
pybit==5.9.0
pandas==2.2.3
river==0.22.0
python-dotenv==1.0.0
numpy==1.24.4
protobuf==3.20.3
tensorflow-probability==0.22.0
ml-dtypes==0.3.2
typing-extensions==4.8.0